Consolidate models and rename constants: merge PlanYaml into PlanModels, rename MenuOrder to Constants, DashboardStats to DashboardModels
