+++ src/Ivy.Tendril.Test/JobServiceCompletionGuardTests.cs
@@ -252,9 +252,9 @@ public class JobServiceCompletionGuardTests
         {
         }
 
-        public DashboardStats GetDashboardData(string? projectFilter)
+        public DashboardModels GetDashboardData(string? projectFilter)
         {
-            return new DashboardStats(0, 0, 0, 0, 0, 0, 0, [], []);
+            return new DashboardModels(0, 0, 0, 0, 0, 0, 0, [], []);
         }
 
         public decimal GetPlanTotalCost(string folderPath)
