+++ src/Ivy.Tendril.Test/JobServiceRetryBlockedTests.cs
@@ -306,9 +306,9 @@ public class JobServiceRetryBlockedTests
         {
         }
 
-        public DashboardStats GetDashboardData(string? projectFilter)
+        public DashboardModels GetDashboardData(string? projectFilter)
         {
-            return new DashboardStats(0, 0, 0, 0, 0, 0, 0, [], []);
+            return new DashboardModels(0, 0, 0, 0, 0, 0, 0, [], []);
         }
 
         public decimal GetPlanTotalCost(string folderPath)
