+++ src/Ivy.Tendril/Apps/DashboardApp.cs
@@ -17,7 +17,7 @@ public class DashboardDayRow
 }
 
 
-[App(title: "Dashboard", icon: Icons.ChartBar, group: ["Apps"], order: MenuOrder.Dashboard)]
+[App(title: "Dashboard", icon: Icons.ChartBar, group: ["Apps"], order: Constants.Dashboard)]
 public class DashboardApp : ViewBase
 {
     public override object Build()
