+++ src/Ivy.Tendril/Apps/OnboardingApp.cs
@@ -4,7 +4,7 @@ using Ivy.Tendril.Models;
 namespace Ivy.Tendril.Apps;
 
 #if DEBUG
-[App(title: "Onboarding", icon: Icons.Rocket, group: ["Debug"], isVisible: true, order: MenuOrder.Onboarding)]
+[App(title: "Onboarding", icon: Icons.Rocket, group: ["Debug"], isVisible: true, order: Constants.Onboarding)]
 #else
 [App(icon: Icons.Rocket, isVisible: false, order: MenuOrder.Onboarding)]
 #endif
