+++ src/Ivy.Tendril/Apps/ReviewApp.cs
@@ -8,7 +8,7 @@ using SidebarView = Ivy.Tendril.Apps.Review.SidebarView;
 
 namespace Ivy.Tendril.Apps;
 
-[App(title: "Review", icon: Icons.ThumbsUp, group: ["Apps"], order: MenuOrder.Review,
+[App(title: "Review", icon: Icons.ThumbsUp, group: ["Apps"], order: Constants.Review,
     allowDuplicateTabs: false)]
 public class ReviewApp : ViewBase
 {
