+++ src/Ivy.Tendril/Services/IPlanDatabaseService.cs
@@ -13,7 +13,7 @@ public interface IPlanDatabaseService : IDisposable
 
     // Aggregates
     PlanReaderService.PlanCountSnapshot ComputePlanCounts();
-    DashboardStats GetDashboardData(string? projectFilter);
+    DashboardModels GetDashboardData(string? projectFilter);
 
     // Costs and tokens
     decimal GetPlanTotalCost(int planId);
